+++ shared/src/commonMain/kotlin/com/toquete/stopclenching/model/AlarmItem.kt
@@ -3,5 +3,5 @@ package com.toquete.stopclenching.model
 data class AlarmItem(
     val from: String,
     val to: String,
-    val intervalMillis: Long
+    val intervalMillis: Int
 )
